[WEB-31] moved the contact page nav bar to the contactpage component and fixed the imports
