+++ src/components/ContactPage/index.ts
@@ -1,3 +1,4 @@
+import ContactPageNavbar from "./ContactPageNavbar";
 import LetsTalkHeadings from "./LetsTalkHeadings";
 
-export { LetsTalkHeadings };
+export { LetsTalkHeadings, ContactPageNavbar };
