+++ src/components/Navbar/index.ts
@@ -10,7 +10,7 @@ import PlatformSubmenu from "./PlatformSubmenu";
 import SideMenuLinks from "./SideMenuLinks";
 import SideMenuLink from "./SideMenuLink";
 import ContactUs from "./ContactUs";
-import ContactPageNavbar from "./ContactPageNavbar";
+
 export {
   Navbar,
   Solutions,
@@ -24,5 +24,4 @@ export {
   SideMenuLinks,
   SideMenuLink,
   ContactUs,
-  ContactPageNavbar,
 };
